+++ src/atoms.tsx
@@ -10,3 +10,8 @@ export const cartSelectAllState = atom<SelectCartItems>({
   key: "isSelectAll",
   default: { id: [], name: [], total_price: 0 },
 });
+
+export const BuyState = atom<boolean>({
+  key: "wasBuyed",
+  default: false,
+});
