--- src/components/Mypage/MyCart/components/Certification.tsx
@@ -1,1 +1,0 @@
-export {};
