--- src/components/Mypage/MyCart/components/CertificationResult.tsx
@@ -1,1 +1,0 @@
-export {};
