--- src/components/Mypage/MyCart/components/PaymentsResult.tsx
@@ -1,1 +1,0 @@
-export {};
